More shuffling for new import dc API
